requirements.txt
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
einops==0.7.0
filelock==3.13.1
ftfy==6.2.0
gmpy2==2.1.2
idna==3.4
imageio==2.34.1
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.3
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
openai-clip==1.0.1
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pip==23.3.1
PySocks==1.7.1
PyWavelets==1.4.1
regex==2024.4.16
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
setuptools==68.2.2
sympy==1.12
terminaltables==3.1.10
tifffile==2023.7.10
torch==2.0.0
torchvision==0.15.0
tqdm==4.66.2
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
wheel==0.41.2
